refactor: Board 폴더로 Notcie, Store Page 이동
