+++ src/pages/BoardPage/Notice/NoticePage/NoticePage.module.scss
