+++ src/pages/BoardPage/Notice/NoticePage/index.tsx
@@ -4,7 +4,7 @@ import HotPost from 'components/Post/HotPost';
 import PostHeader from 'components/Post/PostHeader';
 import PostList from 'components/Post/PostList';
 import LoadingSpinner from 'components/common/LoadingSpinner';
-import useArticleList from 'pages/Notice/hooks/useArticleList';
+import useArticleList from 'pages/BoardPage/Notice/hooks/useArticleList';
 import useParamsHandler from 'utils/hooks/useParamsHandler';
 import styles from './NoticePage.module.scss';
