+++ src/pages/BoardPage/Store/StoreDetailPage/StoreDetailPage.module.scss
